ProjectBook/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-livereload-server==0.4
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
tzdata==2022.5
whitenoise==6.2.0